# contracts/buyer.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PiggiesBuyerService",
    "version": "1.0.0",
    "description": "PROPOSTA contract-first. Implementação ainda não existe. Ver docs/SLP-CONTRATOS.md para decisões e garantias."
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Porta local proposta; precisa ser configurada na implementação."
    }
  ],
  "tags": [
    {
      "name": "Offers"
    }
  ],
  "paths": {
    "/buy-offers": {
      "post": {
        "tags": [
          "Offers"
        ],
        "operationId": "createBuyOffer",
        "summary": "Criar lance de compra de 1 Piggy (processamento assíncrono)",
        "description": "Recebe o comprador e o valor (priceUsd); o servidor gera o offerId e o lance (PENDING), persiste oferta e outbox atomicamente e responde 202 sem aguardar match. Juntar compra e venda é responsabilidade do PiggiesTransactionCoordinator. Cada POST cria um novo lance (sem idempotência por offerId).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOffer"
              },
              "example": {
                "priceUsd": "30.00"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Oferta e evento pendente persistidos; não significa execução concluída.",
            "headers": {
              "Location": {
                "required": true,
                "description": "URL relativa para consultar a oferta",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                },
                "example": "/buy-offers/11111111-1111-4111-8111-111111111111"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedOffer"
                },
                "example": {
                  "offerId": "11111111-1111-4111-8111-111111111111",
                  "status": "PENDING",
                  "statusUrl": "/buy-offers/11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "400": {
            "description": "JSON, UUID ou preço inválido; campos desconhecidos não são aceitos.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "about:blank",
                  "title": "JSON, UUID ou preço inválido; campos desconhecidos não são aceitos.",
                  "status": 400,
                  "detail": "JSON, UUID ou preço inválido; campos desconhecidos não são aceitos.",
                  "code": "INVALID_REQUEST"
                }
              }
            }
          },
          "503": {
            "description": "Aceitação não confirmada; repetir os mesmos dados e offerId. A resposta pode ser incerta após falha de rede.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Aceitação não confirmada; repetir os mesmos dados e offerId. A resposta pode ser incerta após falha de rede.",
                  "status": 503,
                  "detail": "Aceitação não confirmada; repetir os mesmos dados e offerId. A resposta pode ser incerta após falha de rede.",
                  "code": "SERVICE_UNAVAILABLE"
                }
              }
            }
          }
        }
      }
    },
    "/buy-offers/{offerId}": {
      "get": {
        "tags": [
          "Offers"
        ],
        "operationId": "getBuyOffer",
        "summary": "Consultar oferta e resultado eventual",
        "description": "PENDING inclui evento pendente de publicação e oferta aguardando match. Não há prazo garantido para match. Estados terminais não regridem. Não recebe body.",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado atual; atualização por eventos é eventualmente consistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Offer"
                },
                "example": {
                  "offerId": "11111111-1111-4111-8111-111111111111",
                  "buyerId": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "priceUsd": "30.00",
                  "createdAt": "2026-09-28T18:00:00Z",
                  "status": "EXECUTED",
                  "transactionId": "33333333-3333-4333-8333-333333333333",
                  "executionPriceUsd": "25.000"
                }
              }
            }
          },
          "400": {
            "description": "JSON, UUID ou preço inválido; campos desconhecidos não são aceitos.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "about:blank",
                  "title": "JSON, UUID ou preço inválido; campos desconhecidos não são aceitos.",
                  "status": 400,
                  "detail": "JSON, UUID ou preço inválido; campos desconhecidos não são aceitos.",
                  "code": "INVALID_REQUEST"
                }
              }
            }
          },
          "404": {
            "description": "Oferta inexistente.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Oferta inexistente.",
                  "status": 404,
                  "detail": "Oferta inexistente.",
                  "code": "OFFER_NOT_FOUND"
                }
              }
            }
          },
          "503": {
            "description": "Consulta temporariamente indisponível.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Consulta temporariamente indisponível.",
                  "status": 503,
                  "detail": "Consulta temporariamente indisponível.",
                  "code": "SERVICE_UNAVAILABLE"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateOffer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "buyerId": {
            "type": "string",
            "format": "uuid",
            "description": "Opcional (MVP). Se ausente, o servidor gera."
          },
          "priceUsd": {
            "type": "string",
            "pattern": "^(?!0(?:\\.0{1,2})?$)(?:0|[1-9][0-9]{0,11})(?:\\.[0-9]{1,2})?$",
            "description": "USD positivo; até 12 dígitos inteiros e 2 casas decimais; sem expoente. Converter com BigDecimal."
          }
        },
        "required": [
          "priceUsd"
        ]
      },
      "AcceptedOffer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "const": "PENDING"
          },
          "statusUrl": {
            "type": "string",
            "format": "uri-reference",
            "pattern": "^/buy-offers/[0-9a-fA-F-]{36}$"
          }
        },
        "required": [
          "offerId",
          "status",
          "statusUrl"
        ]
      },
      "Offer": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PendingOffer"
          },
          {
            "$ref": "#/components/schemas/ExecutedOffer"
          }
        ]
      },
      "PendingOffer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "buyerId": {
            "type": "string",
            "format": "uuid"
          },
          "priceUsd": {
            "type": "string",
            "pattern": "^(?!0(?:\\.0{1,2})?$)(?:0|[1-9][0-9]{0,11})(?:\\.[0-9]{1,2})?$",
            "description": "USD positivo; até 12 dígitos inteiros e 2 casas decimais; sem expoente. Converter com BigDecimal."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "const": "PENDING"
          }
        },
        "required": [
          "offerId",
          "buyerId",
          "priceUsd",
          "createdAt",
          "status"
        ]
      },
      "ExecutedOffer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "buyerId": {
            "type": "string",
            "format": "uuid"
          },
          "priceUsd": {
            "type": "string",
            "pattern": "^(?!0(?:\\.0{1,2})?$)(?:0|[1-9][0-9]{0,11})(?:\\.[0-9]{1,2})?$",
            "description": "USD positivo; até 12 dígitos inteiros e 2 casas decimais; sem expoente. Converter com BigDecimal."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "const": "EXECUTED"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "executionPriceUsd": {
            "type": "string",
            "pattern": "^(?!0(?:\\.0{1,3})?$)(?:0|[1-9][0-9]{0,11})(?:\\.[0-9]{1,3})?$",
            "description": "USD positivo; até 12 dígitos inteiros e 3 casas decimais; sem expoente. Converter com BigDecimal."
          }
        },
        "required": [
          "offerId",
          "buyerId",
          "priceUsd",
          "createdAt",
          "status",
          "transactionId",
          "executionPriceUsd"
        ]
      },
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ]
      }
    }
  }
}